Removed data files, and cleaned up loading in phaser games
