+++ app/writing/the-sun/page.tsx
@@ -1,8 +1,45 @@
 import StoryContentLoader from "@/src/components/StoryContentLoader";
-import { theSunData } from "@/src/data/writing/the-sun-data";
+import { StoryDataProps } from "@/src/types/data-props";
 import ResumeScrollProgress from "@/src/components/ResumeScrollProgress";
 
-const storyData = theSunData;
+const storyData: StoryDataProps = {
+  title: "The Sun",
+  subtitle: "Shane Bonkowski",
+  date: "August 25, 2024",
+  imageUrl: "/webps/writing/the-sun.webp",
+  imageWidth: 500,
+  imageHeight: 422,
+  body: [
+    {
+      content: `To my firstborn,
+
+Do you ever stop and wonder, if the wind did not blow, would the wilted flowers remain in frozen perfection for all eternity? Suspended in time like a delicate hourglass whose sand does not trickle. Yet, would the flower’s beauty be as profound if it did not softly surrender to the wind’s embrace? 
+
+Spread your mighty wings and join me on this journey through paradise. Stay close by my side. Where we go the light will follow. 
+
+Down below, the forest is coming to life. The songbirds spin their beautiful melodies as my light peeks through the trees. The hummingbirds drink the sweet nectar from the fresh flowers and spread their pollen across the fields. Among the treetops, the crows glide silently between the branches. Their dark feathers contrast against my pale light filtering through the leaves like black stars, bringing with them the passion of the night sky. In the symphony that fills the forest, they are the occasional booming call of the war drum. For a brief moment, the forest holds its breath, silent and attentive, not resuming until the drum pounds again. 
+
+I admire the crows.
+
+All across the forest, the birds are hard at work nesting and finding mates. The long winter has finally passed, and the young and ambitious are ready to start their families. Some spend most of their days fortifying their nests, having seen the horrors that late spring can bring. Others spend most of their time singing their lustful songs and shimmering with their colorful displays, eager to win over a mate. As spring grows older, the nests spread from treetop to treetop with a canopy of birds. Some are thin and frail, like delicate spiderwebs caught in the breeze. Others are fortresses, built to last. 
+
+When the storms of late spring tear through this forest, the mothers will spread their nurturing wings and clutch their delicate eggs. In some nests, like the crows, the fathers will stretch their monolithic wings, clinging to their partners and the future they protect together. The rain will pour down from the sky like the oceans that cover this beautiful realm, pelting the birds like artillery shells. The same wind they use to soar among the clouds will force its prying hands upon them, trying to rip them from their fortress. Lightning will strike down from the heavens and fires will rage in its wake.
+
+Those with weaker nests will have already fled for safety, while the crows will endure the storm for as long as it takes. They will shield their incubating young from the onslaught above, remaining as still as statues, unshaken by the winds that aim to fell their fortress. 
+
+And though not all can weather the storm, their presence lingers on in the fortresses they have left behind, sheltering their families so they can carry on. 
+
+And when the storms finally pass and their light creeps out from behind the clouds, their families will rise victorious and soar through the skies on the wind born from the warmth they provide. 
+
+And they will rebuild.
+
+Until we meet again,
+- The light that beats upon your back as you soar among the treetops.`,
+      fontStyle: "normal",
+      textAlign: "justify",
+    },
+  ],
+};
 
 export const metadata = {
   title: storyData.title,
