--- src/data/README.md
@@ -1,13 +1,0 @@
-## Data
-
-This directory is used to store all the data that need to be processed to create the pages on the website. This directory primarily contains the content for various pages of the application, including the writing pages, the about me page, and other text-heavy pages, as well as the metadata for the games.
-
-### Contents
-
-- JSON Files
-- JavaScript/TypeScript Files
-- Text Content
-
-### Usage
-
-The data stored in this directory is typically imported and processed by components or pages in the application. This approach helps in separating content from presentation, making the codebase more modular and maintainable.
