--- src/data/art/strange-love-data.ts
@@ -1,19 +1,0 @@
-import { StoryDataProps } from "@/src/types/data-props";
-
-export const strangeLoveData: StoryDataProps = {
-  title: "Strange Love",
-  subtitle: "Shane Bonkowski",
-  date: "April 6, 2022",
-  imageUrl: "/webps/art/strange-love.webp",
-  imageWidth: 1194,
-  imageHeight: 834,
-  artContent: true,
-  body: [
-    {
-      content:
-        "Bone of my bones, flesh of my flesh. A piece of me, now a piece of you.",
-      fontStyle: "italic",
-      textAlign: "center",
-    },
-  ],
-};
