--- src/data/games/better-boids-data.ts
@@ -1,8 +1,0 @@
-import { GameDataProps } from "@/src/types/data-props";
-
-export const betterBoidsData: GameDataProps = {
-  title: "Better Boids",
-  description: "A game by Shane Bonkowski.",
-  logoImageUrl: "/webps/mars-logo-large.webp",
-  imageAlt: "Better Boids",
-};
