--- src/data/games/game-template-data.ts
@@ -1,8 +1,0 @@
-import { GameDataProps } from "@/src/types/data-props";
-
-export const gameTemplateData: GameDataProps = {
-  title: "Game Template",
-  description: "A game by Shane Bonkowski.",
-  logoImageUrl: "/webps/mars-logo-large.webp",
-  imageAlt: "Game Template",
-};
