--- src/data/main/cookie-policy-data.ts
@@ -1,172 +1,0 @@
-import { ContentDataProps } from "@/src/types/data-props";
-
-export const cookiePolicyData: ContentDataProps[] = [
-  {
-    type: "h1",
-    text: "Cookie Policy",
-  },
-  {
-    type: "paragraph",
-    text: 'This Cookie Policy explains how Shanes Games ("Company," "We," "Us," and "Our") uses cookies and similar technologies to recognize you when you visit our website at <a href="https://shanebonkowski.com">https://shanebonkowski.com</a> ("Website"). It explains what these technologies are and why we use them, as well as your rights to control our use of them.',
-  },
-  {
-    type: "paragraph",
-    text: "In some cases we may use cookies to collect personal information, or that becomes personal information if we combine it with other information.",
-  },
-  {
-    type: "paragraph",
-    text: "Disclaimer: the following document outlines the general practices and policies regarding cookies. Not all mentioned scenarios, cookies, or processes may be currently implemented, but could be in the future.",
-  },
-  {
-    type: "h2",
-    text: "What are cookies?",
-  },
-  {
-    type: "paragraph",
-    text: "Cookies are small data files that are placed on your computer or mobile device when you visit a website. Cookies are widely used by website owners in order to make their websites work, or to work more efficiently, as well as to provide reporting information.",
-  },
-  {
-    type: "paragraph",
-    text: 'Cookies set by the website owner (in this case, Shanes Games) are called "first-party cookies." Cookies set by parties other than the website owner are called "third-party cookies." Third-party cookies enable third-party features or functionality to be provided on or through the website (e.g., advertising, interactive content, and analytics). The parties that set these third-party cookies can recognize your computer both when it visits the website in question and also when it visits certain other websites.',
-  },
-  {
-    type: "h2",
-    text: "Why do we use cookies?",
-  },
-  {
-    type: "paragraph",
-    text: 'We use first- and third-party cookies for several reasons. Some cookies are required for technical reasons in order for our Website to operate, and we refer to these as "essential" or "strictly necessary" cookies. Other cookies also enable us to track and target the interests of our users to enhance the experience on our Online Properties. Third parties serve cookies through our Website for advertising, analytics, and other purposes. This is described in more detail below.',
-  },
-  {
-    type: "h2",
-    text: "How can I control cookies?",
-  },
-  {
-    type: "paragraph",
-    text: "You have the right to decide whether to accept or reject cookies. Essential cookies cannot be rejected as they are strictly necessary to provide you with services.",
-  },
-  {
-    type: "paragraph",
-    text: "If you choose to reject cookies, you may still use our website though your access to some functionality and areas of our website may be restricted. You may also set or amend your web browser controls to accept or refuse cookies.",
-  },
-  {
-    type: "paragraph",
-    text: "The specific types of first- and third-party cookies served through our Website and the purposes they perform are described in the table below (please note that the specific cookies served may vary depending on the specific Online Properties you visit):",
-  },
-  {
-    type: "h3",
-    text: "Necessary / Essential Cookies",
-  },
-  {
-    type: "list",
-    items: [
-      "Type: Session Cookies",
-      "Administered by: Us",
-      "Purpose: These Cookies are essential to provide You with services available through the Website and to enable You to use some of its features. They help to authenticate users and prevent fraudulent use of user accounts. Without these Cookies, the services that You have asked for cannot be provided, and We only use these Cookies to provide You with those services.",
-    ],
-  },
-  {
-    type: "h3",
-    text: "Cookies Policy / Notice Acceptance Cookies",
-  },
-  {
-    type: "list",
-    items: [
-      "Type: Persistent Cookies",
-      "Administered by: Us",
-      "Purpose: These Cookies identify if users have accepted the use of cookies on the Website.",
-    ],
-  },
-  {
-    type: "h3",
-    text: "Functionality Cookies",
-  },
-  {
-    type: "list",
-    items: [
-      "Type: Persistent Cookies",
-      "Administered by: Us",
-      "Purpose: These Cookies allow us to remember choices You make when You use the Website, such as remembering your login details or language preference. The purpose of these Cookies is to provide You with a more personal experience and to avoid You having to re-enter your preferences every time You use the Website.",
-    ],
-  },
-  {
-    type: "h3",
-    text: "Tracking and Performance Cookies",
-  },
-  {
-    type: "list",
-    items: [
-      "Type: Persistent Cookies",
-      "Administered by: Third-Parties",
-      "Purpose: These Cookies are used to track information about traffic to the Website and how users use the Website. The information gathered via these Cookies may directly or indirectly identify you as an individual visitor. This is because the information collected is typically linked to a pseudonymous identifier associated with the device you use to access the Website. We may also use these Cookies to test new pages, features or new functionality of the Website to see how our users react to them.",
-    ],
-  },
-  {
-    type: "h3",
-    text: "Targeting and Advertising Cookies",
-  },
-  {
-    type: "list",
-    items: [
-      "Type: Persistent Cookies",
-      "Administered by: Third-Parties",
-      "Purpose: These Cookies track your browsing habits to enable Us to show advertising which is more likely to be of interest to You. These Cookies use information about your browsing history to group You with other users who have similar interests. Based on that information, and with Our permission, third party advertisers can place Cookies to enable them to show adverts which We think will be relevant to your interests while You are on third party websites.",
-    ],
-  },
-  {
-    type: "h2",
-    text: "How can I control cookies on my browser?",
-  },
-  {
-    type: "paragraph",
-    text: "As the means by which you can refuse cookies through your web browser controls vary from browser to browser, you should visit your browsers help menu for more information.",
-  },
-  {
-    type: "paragraph",
-    text: "In addition, most advertising networks offer you a way to opt out of targeted advertising.",
-  },
-  {
-    type: "h2",
-    text: "What about other tracking technologies, like web beacons?",
-  },
-  {
-    type: "paragraph",
-    text: 'Cookies are not the only way to recognize or track visitors to a website. We may use other, similar technologies from time to time, like web beacons (sometimes called "tracking pixels" or "clear gifs"). These are tiny graphics files that contain a unique identifier that enables us to recognize when someone has visited our Website or opened an email including them. This allows us, for example, to monitor the traffic patterns of users from one page within a website to another, to deliver or communicate with cookies, to understand whether you have come to the website from an online advertisement displayed on a third-party website, to improve site performance, and to measure the success of email marketing campaigns. In many instances, these technologies are reliant on cookies to function properly, and so declining cookies will impair their functioning.',
-  },
-  {
-    type: "h2",
-    text: "Do you use Flash cookies or Local Shared Objects?",
-  },
-  {
-    type: "paragraph",
-    text: 'Websites may also use so-called "Flash Cookies" (also known as Local Shared Objects or "LSOs") to, among other things, collect and store information about your use of our services, fraud prevention, and for other site operations.',
-  },
-  {
-    type: "paragraph",
-    text: 'If you do not want Flash Cookies stored on your computer, you can adjust the settings of your Flash player to block Flash Cookies storage using the tools contained in the Website Storage Settings Panel. You can also control Flash Cookies by going to the Global Storage Settings Panel and following the instructions (which may include instructions that explain, for example, how to delete existing Flash Cookies (referred to "information" on the Macromedia site), how to prevent Flash LSOs from being placed on your computer without your being asked, and (for Flash Player 8 and later) how to block Flash Cookies that are not being delivered by the operator of the page you are on at the time).',
-  },
-  {
-    type: "paragraph",
-    text: "Please note that setting the Flash Player to restrict or limit acceptance of Flash Cookies may reduce or impede the functionality of some Flash applications, including, potentially, Flash applications used in connection with our services or online content.",
-  },
-  {
-    type: "h2",
-    text: "Do you serve targeted advertising?",
-  },
-  {
-    type: "paragraph",
-    text: "Third parties may serve cookies on your computer or mobile device to serve advertising through our Website. These companies may use information about your visits to this and other websites in order to provide relevant advertisements about goods and services that you may be interested in. They may also employ technology that is used to measure the effectiveness of advertisements. They can accomplish this by using cookies or web beacons to collect information about your visits to this and other sites in order to provide relevant advertisements about goods and services of potential interest to you. The information collected through this process does not enable us or them to identify your name, contact details, or other details that directly identify you unless you choose to provide these.",
-  },
-  {
-    type: "h2",
-    text: "Contact Us",
-  },
-  {
-    type: "paragraph",
-    text: "If you have any questions about this Cookie Policy, You can contact us:",
-  },
-  {
-    type: "paragraph",
-    text: "By email: 3ravens3@gmail.com",
-  },
-];
