--- src/data/main/not-found-data.ts
@@ -1,12 +1,0 @@
-import { ContentDataProps } from "@/src/types/data-props";
-
-export const notFoundData: ContentDataProps[] = [
-  {
-    type: "h1",
-    text: "404 - Page Not Found",
-  },
-  {
-    type: "paragraph",
-    text: "Sorry, the page you are looking for does not exist.",
-  },
-];
