--- src/data/main/privacy-policy-data.ts
@@ -1,492 +1,0 @@
-import { ContentDataProps } from "@/src/types/data-props";
-
-export const privacyPolicyData: ContentDataProps[] = [
-  {
-    type: "h1",
-    text: "Privacy Policy",
-  },
-  {
-    type: "paragraph",
-    text: "This Privacy Policy describes Our policies and procedures on the collection, use and disclosure of Your information when You use the Service and tells You about Your privacy rights and how the law protects You.",
-  },
-  {
-    type: "paragraph",
-    text: "Disclaimer: the following document outlines the general practices and policies regarding data handling and privacy. Not all mentioned scenarios or processes may be currently implemented, but could be in the future.",
-  },
-  {
-    type: "paragraph",
-    text: "We use Your Personal data to provide and improve the Service. By using the Service, You agree to the collection and use of information in accordance with this Privacy Policy.",
-  },
-  {
-    type: "h2",
-    text: "Interpretation and Definitions",
-  },
-  {
-    type: "h3",
-    text: "Interpretation",
-  },
-  {
-    type: "paragraph",
-    text: "The words of which the initial letter is capitalized have meanings defined under the following conditions. The following definitions shall have the same meaning regardless of whether they appear in singular or in plural.",
-  },
-  {
-    type: "h3",
-    text: "Definitions",
-  },
-  {
-    type: "paragraph",
-    text: "For the purposes of this Privacy Policy:",
-  },
-  {
-    type: "list",
-    items: [
-      "Account means a unique account created for You to access our Service or parts of our Service.",
-      'Affiliate means an entity that controls, is controlled by or is under common control with a party, where "control" means ownership of 50% or more of the shares, equity interest or other securities entitled to vote for election of directors or other managing authority.',
-      "Application refers to Better Boids, Flip Tile, and any other game contained on this Website.",
-      'Company (referred to as either "the Company", "We", "Us" or "Our" in this Agreement) refers to Shanes Games.',
-      "For the purpose of the GDPR, the Company is the Data Controller.",
-      "Cookies are small files that are placed on Your computer, mobile device or any other device by a website, containing the details of Your browsing history on that website among its many uses.",
-      "Country refers to: California, United States",
-      "Data Controller, for the purposes of the GDPR (General Data Protection Regulation), refers to the Company as the legal person which alone or jointly with others determines the purposes and means of the processing of Personal Data.",
-      "Device means any device that can access the Service such as a computer, a cellphone or a digital tablet.",
-      "GDPR refers to EU General Data Protection Regulation.",
-      "Personal Data is any information that relates to an identified or identifiable individual.",
-      "For the purposes of GDPR, Personal Data means any information relating to You such as a name, an identification number, location data, online identifier or to one or more factors specific to the physical, physiological, genetic, mental, economic, cultural or social identity.",
-      "Service refers to the Application or the Website or both.",
-      "Service Provider means any natural or legal person who processes the data on behalf of the Company. It refers to third-party companies or individuals employed by the Company to facilitate the Service, to provide the Service on behalf of the Company, to perform services related to the Service or to assist the Company in analyzing how the Service is used. For the purpose of the GDPR, Service Providers are considered Data Processors.",
-      "Usage Data refers to data collected automatically, either generated by the use of the Service or from the Service infrastructure itself (for example, the duration of a page visit).",
-      "Website refers to <a href='/'>Shanes Games</a>.",
-      "You means the individual accessing or using the Service, or the company, or other legal entity on behalf of which such individual is accessing or using the Service, as applicable.",
-      "Under GDPR, You can be referred to as the Data Subject or as the User as you are the individual using the Service.",
-    ],
-  },
-  {
-    type: "h2",
-    text: "Collecting and Using Your Personal Data",
-  },
-  {
-    type: "h3",
-    text: "Types of Data Collected",
-  },
-  {
-    type: "h3",
-    text: "Personal Data",
-  },
-  {
-    type: "paragraph",
-    text: "While using Our Service, We may ask You to provide Us with certain personally identifiable information that can be used to contact or identify You. Personally identifiable information may include, but is not limited to:",
-  },
-  {
-    type: "h3",
-    text: "Usage Data",
-  },
-  {
-    type: "paragraph",
-    text: "Usage Data is collected automatically when using the Service.",
-  },
-  {
-    type: "paragraph",
-    text: "Usage Data may include information such as Your Device's Internet Protocol address (e.g. IP address), browser type, browser version, the pages of our Service that You visit, the time and date of Your visit, the time spent on those pages, unique device identifiers and other diagnostic data.",
-  },
-  {
-    type: "paragraph",
-    text: "When You access the Service by or through a mobile device, We may collect certain information automatically, including, but not limited to, the type of mobile device You use, Your mobile device unique ID, the IP address of Your mobile device, Your mobile operating system, the type of mobile Internet browser You use, unique device identifiers and other diagnostic data.",
-  },
-  {
-    type: "paragraph",
-    text: "We may also collect information that Your browser sends whenever You visit our Service or when You access the Service by or through a mobile device.",
-  },
-  {
-    type: "h3",
-    text: "Tracking Technologies and Cookies",
-  },
-  {
-    type: "paragraph",
-    text: "We use Cookies and similar tracking technologies to track the activity on Our Service and store certain information. Tracking technologies used are beacons, tags, and scripts to collect and track information and to improve and analyze Our Service. The technologies We use may include:",
-  },
-  {
-    type: "list",
-    items: [
-      "Cookies or Browser Cookies. A cookie is a small file placed on Your Device. You can instruct Your browser to refuse all Cookies or to indicate when a Cookie is being sent. However, if You do not accept Cookies, You may not be able to use some parts of our Service. Unless you have adjusted Your browser setting so that it will refuse Cookies, our Service may use Cookies.",
-      "Web Beacons. Certain sections of our Service and our emails may contain small electronic files known as web beacons (also referred to as clear gifs, pixel tags, and single-pixel gifs) that permit the Company, for example, to count users who have visited those pages or opened an email and for other related website statistics (for example, recording the popularity of a certain section and verifying system and server integrity).",
-      'Cookies can be "Persistent" or "Session" Cookies. Persistent Cookies remain on Your personal computer or mobile device when You go offline, while Session Cookies are deleted as soon as You close Your web browser.',
-    ],
-  },
-  {
-    type: "paragraph",
-    text: "We use both Session and Persistent Cookies for the purposes set out below:",
-  },
-  {
-    type: "h3",
-    text: "Necessary / Essential Cookies",
-  },
-  {
-    type: "list",
-    items: [
-      "Type: Session Cookies",
-      "Administered by: Us",
-      "Purpose: These Cookies are essential to provide You with services available through the Website and to enable You to use some of its features. They help to authenticate users and prevent fraudulent use of user accounts. Without these Cookies, the services that You have asked for cannot be provided, and We only use these Cookies to provide You with those services.",
-    ],
-  },
-  {
-    type: "h3",
-    text: "Cookies Policy / Notice Acceptance Cookies",
-  },
-  {
-    type: "list",
-    items: [
-      "Type: Persistent Cookies",
-      "Administered by: Us",
-      "Purpose: These Cookies identify if users have accepted the use of cookies on the Website.",
-    ],
-  },
-  {
-    type: "h3",
-    text: "Functionality Cookies",
-  },
-  {
-    type: "list",
-    items: [
-      "Type: Persistent Cookies",
-      "Administered by: Us",
-      "Purpose: These Cookies allow us to remember choices You make when You use the Website, such as remembering your login details or language preference. The purpose of these Cookies is to provide You with a more personal experience and to avoid You having to re-enter your preferences every time You use the Website.",
-    ],
-  },
-  {
-    type: "h3",
-    text: "Tracking and Performance Cookies",
-  },
-  {
-    type: "list",
-    items: [
-      "Type: Persistent Cookies",
-      "Administered by: Third-Parties",
-      "Purpose: These Cookies are used to track information about traffic to the Website and how users use the Website. The information gathered via these Cookies may directly or indirectly identify you as an individual visitor. This is because the information collected is typically linked to a pseudonymous identifier associated with the device you use to access the Website. We may also use these Cookies to test new pages, features or new functionality of the Website to see how our users react to them.",
-    ],
-  },
-  {
-    type: "h3",
-    text: "Targeting and Advertising Cookies",
-  },
-  {
-    type: "list",
-    items: [
-      "Type: Persistent Cookies",
-      "Administered by: Third-Parties",
-      "Purpose: These Cookies track your browsing habits to enable Us to show advertising which is more likely to be of interest to You. These Cookies use information about your browsing history to group You with other users who have similar interests. Based on that information, and with Our permission, third party advertisers can place Cookies to enable them to show adverts which We think will be relevant to your interests while You are on third party websites.",
-    ],
-  },
-  {
-    type: "paragraph",
-    text: "For more information about the cookies we use and your choices regarding cookies, please visit our <a href='/main/cookie-policy'>Cookies Policy</a> or the Cookies section of our Privacy Policy.",
-  },
-  {
-    type: "h2",
-    text: "Use of Your Personal Data",
-  },
-  {
-    type: "paragraph",
-    text: "The Company may use Personal Data for the following purposes:",
-  },
-  {
-    type: "list",
-    items: [
-      "To provide and maintain our Service, including to monitor the usage of our Service.",
-      "To manage Your Account: to manage Your registration as a user of the Service. The Personal Data You provide can give You access to different functionalities of the Service that are available to You as a registered user.",
-      "For the performance of a contract: the development, compliance and undertaking of the purchase contract for the products, items or services You have purchased or of any other contract with Us through the Service.",
-      "To contact You: To contact You by email, telephone calls, SMS, or other equivalent forms of electronic communication, such as a mobile application's push notifications regarding updates or informative communications related to the functionalities, products or contracted services, including the security updates, when necessary or reasonable for their implementation.",
-      "To provide You with news, special offers and general information about other goods, services and events which we offer that are similar to those that you have already purchased or enquired about unless You have opted not to receive such information.",
-      "To manage Your requests: To attend and manage Your requests to Us.",
-      "For business transfers: We may use Your information to evaluate or conduct a merger, divestiture, restructuring, reorganization, dissolution, or other sale or transfer of some or all of Our assets, whether as a going concern or as part of bankruptcy, liquidation, or similar proceeding, in which Personal Data held by Us about our Service users is among the assets transferred.",
-      "For other purposes: We may use Your information for other purposes, such as data analysis, identifying usage trends, determining the effectiveness of our promotional campaigns and to evaluate and improve our Service, products, services, marketing and your experience.",
-    ],
-  },
-  {
-    type: "paragraph",
-    text: "We may share Your personal information in the following situations:",
-  },
-  {
-    type: "list",
-    items: [
-      "With Service Providers: We may share Your personal information with Service Providers to monitor and analyze the use of our Service, to show advertisements to You to help support and maintain Our Service, to contact You.",
-      "For business transfers: We may share or transfer Your personal information in connection with, or during negotiations of, any merger, sale of Company assets, financing, or acquisition of all or a portion of Our business to another company.",
-      "With Affiliates: We may share Your information with Our affiliates, in which case we will require those affiliates to honor this Privacy Policy. Affiliates include Our parent company and any other subsidiaries, joint venture partners or other companies that We control or that are under common control with Us.",
-      "With business partners: We may share Your information with Our business partners to offer You certain products, services or promotions.- With other users: when You share personal information or otherwise interact in the public areas with other users, such information may be viewed by all users and may be publicly distributed outside.",
-      "With Your consent: We may disclose Your personal information for any other purpose with Your consent.",
-    ],
-  },
-  {
-    type: "h2",
-    text: "Retention of Your Personal Data",
-  },
-  {
-    type: "paragraph",
-    text: "The Company will retain Your Personal Data only for as long as is necessary for the purposes set out in this Privacy Policy. We will retain and use Your Personal Data to the extent necessary to comply with our legal obligations (for example, if we are required to retain your data to comply with applicable laws), resolve disputes, and enforce our legal agreements and policies.",
-  },
-  {
-    type: "paragraph",
-    text: "The Company will also retain Usage Data for internal analysis purposes. Usage Data is generally retained for a shorter period of time, except when this data is used to strengthen the security or to improve the functionality of Our Service, or We are legally obligated to retain this data for longer time periods.",
-  },
-  {
-    type: "h2",
-    text: "Transfer of Your Personal Data",
-  },
-  {
-    type: "paragraph",
-    text: "Your information, including Personal Data, is processed at the Company's operating offices and in any other places where the parties involved in the processing are located. It means that this information may be transferred to â€” and maintained on â€” computers located outside of Your state, province, country or other governmental jurisdiction where the data protection laws may differ than those from Your jurisdiction.",
-  },
-  {
-    type: "paragraph",
-    text: "Your consent to this Privacy Policy followed by Your submission of such information represents Your agreement to that transfer.",
-  },
-  {
-    type: "paragraph",
-    text: "The Company will take all steps reasonably necessary to ensure that Your data is treated securely and in accordance with this Privacy Policy and no transfer of Your Personal Data will take place to an organization or a country unless there are adequate controls in place including the security of Your data and other personal information.",
-  },
-  {
-    type: "h2",
-    text: "Delete Your Personal Data",
-  },
-  {
-    type: "paragraph",
-    text: "You have the right to delete or request that We assist in deleting the Personal Data that We have collected about You.",
-  },
-  {
-    type: "paragraph",
-    text: "Our Service may give You the ability to delete certain information about You from within the Service.",
-  },
-  {
-    type: "paragraph",
-    text: "You may update, amend, or delete Your information at any time by signing in to Your Account, if you have one, and visiting the account settings section that allows you to manage Your personal information. You may also contact Us to request access to, correct, or delete any personal information that You have provided to Us.",
-  },
-  {
-    type: "paragraph",
-    text: "Please note, however, that We may need to retain certain information when we have a legal obligation or lawful basis to do so.",
-  },
-  {
-    type: "h2",
-    text: "Disclosure of Your Personal Data",
-  },
-  {
-    type: "h3",
-    text: "Business Transactions",
-  },
-  {
-    type: "paragraph",
-    text: "If the Company is involved in a merger, acquisition or asset sale, Your Personal Data may be transferred. We will provide notice before Your Personal Data is transferred and becomes subject to a different Privacy Policy.",
-  },
-  {
-    type: "h3",
-    text: "Law enforcement",
-  },
-  {
-    type: "paragraph",
-    text: "Under certain circumstances, the Company may be required to disclose Your Personal Data if required to do so by law or in response to valid requests by public authorities (e.g. a court or a government agency).",
-  },
-  {
-    type: "h3",
-    text: "Other legal requirements",
-  },
-  {
-    type: "paragraph",
-    text: "The Company may disclose Your Personal Data in the good faith belief that such action is necessary to:",
-  },
-  {
-    type: "h3",
-    text: "Comply with a legal obligation",
-  },
-  {
-    type: "list",
-    items: [
-      "Protect and defend the rights or property of the Company",
-      "Prevent or investigate possible wrongdoing in connection with the Service",
-      "Protect the personal safety of Users of the Service or the public",
-      "Protect against legal liability",
-      "Security of Your Personal Data",
-      "The security of Your Personal Data is important to Us, but remember that no method of transmission over the Internet, or method of electronic storage is 100% secure. While We strive to use commercially acceptable means to protect Your Personal Data, We cannot guarantee its absolute security.",
-    ],
-  },
-  {
-    type: "h2",
-    text: "Detailed Information on the Processing of Your Personal Data",
-  },
-  {
-    type: "paragraph",
-    text: "The Service Providers We use may have access to Your Personal Data. These third-party vendors collect, store, use, process and transfer information about Your activity on Our Service in accordance with their Privacy Policies.",
-  },
-  {
-    type: "h3",
-    text: "Analytics",
-  },
-  {
-    type: "paragraph",
-    text: "We may use third-party Service providers to monitor and analyze the use of our Service.",
-  },
-  {
-    type: "h2",
-    text: "Google Analytics",
-  },
-  {
-    type: "paragraph",
-    text: "Google Analytics is a web analytics service offered by Google that tracks and reports website traffic. Google uses the data collected to track and monitor the use of our Service. This data is shared with other Google services. Google may use the collected data to contextualize and personalize the ads of its own advertising network.",
-  },
-  {
-    type: "paragraph",
-    text: "You can opt-out of having made your activity on the Service available to Google Analytics by installing the Google Analytics opt-out browser add-on. The add-on prevents the Google Analytics JavaScript (ga.js, analytics.js and dc.js) from sharing information with Google Analytics about visits activity.",
-  },
-  {
-    type: "paragraph",
-    text: "You may opt-out of certain Google Analytics features through your mobile device settings, such as your device advertising settings or by following the instructions provided by Google in their Privacy Policy: <a href='https://policies.google.com/privacy' target='_blank' rel='noopener noreferrer'>Google Privacy Policies</a>",
-  },
-  {
-    type: "paragraph",
-    text: "For more information on the privacy practices of Google, please visit the Google Privacy & Terms web page: <a href='https://policies.google.com/privacy' target='_blank' rel='noopener noreferrer'>Google Privacy Policies</a>",
-  },
-  {
-    type: "h2",
-    text: "Advertising",
-  },
-  {
-    type: "paragraph",
-    text: "We may use Service Providers to show advertisements to You to help support and maintain Our Service.",
-  },
-  {
-    type: "h3",
-    text: "Google AdSense & DoubleClick Cookie",
-  },
-  {
-    type: "paragraph",
-    text: "Google, as a third party vendor, uses cookies to serve ads on our Service. Google's use of the DoubleClick cookie enables it and its partners to serve ads to our users based on their visit to our Service or other websites on the Internet.",
-  },
-  {
-    type: "paragraph",
-    text: "You may opt out of the use of the DoubleClick Cookie for interest-based advertising by visiting the Google Ads Settings web page: <a href='https://www.google.com/ads/preferences/' target='_blank' rel='noopener noreferrer'>Google Ads Preferences</a>",
-  },
-  {
-    type: "h3",
-    text: "AdMob by Google",
-  },
-  {
-    type: "paragraph",
-    text: "AdMob by Google is provided by Google Inc.",
-  },
-  {
-    type: "paragraph",
-    text: "You can opt-out from the AdMob by Google service by following the instructions described by Google: <a href='https://support.google.com/ads/answer/2662922?hl=en' target='_blank' rel='noopener noreferrer'>AdMob Opt Out</a>",
-  },
-  {
-    type: "paragraph",
-    text: 'For more information on how Google uses the collected information, please visit the "How Google uses data when you use our partners sites or app" page: <a href="https://policies.google.com/technologies/partner-sites" target="_blank" rel="noopener noreferrer">Google Partner Sites Policies</a> or visit the Privacy Policy of Google: <a href="https://policies.google.com/privacy" target="_blank" rel="noopener noreferrer">Google Privacy Policies</a>',
-  },
-  {
-    type: "h2",
-    text: "GDPR Privacy",
-  },
-  {
-    type: "h3",
-    text: "Legal Basis for Processing Personal Data under GDPR",
-  },
-  {
-    type: "paragraph",
-    text: "We may process Personal Data under the following conditions:",
-  },
-  {
-    type: "list",
-    items: [
-      "Consent: You have given Your consent for processing Personal Data for one or more specific purposes.",
-      "Performance of a contract: Provision of Personal Data is necessary for the performance of an agreement with You and/or for any pre-contractual obligations thereof.",
-      "Legal obligations: Processing Personal Data is necessary for compliance with a legal obligation to which the Company is subject.",
-      "Vital interests: Processing Personal Data is necessary in order to protect Your vital interests or of another natural person.",
-      "Public interests: Processing Personal Data is related to a task that is carried out in the public interest or in the exercise of official authority vested in the Company.",
-      "Legitimate interests: Processing Personal Data is necessary for the purposes of the legitimate interests pursued by the Company.",
-      "In any case, the Company will gladly help to clarify the specific legal basis that applies to the processing, and in particular whether the provision of Personal Data is a statutory or contractual requirement, or a requirement necessary to enter into a contract.",
-    ],
-  },
-  {
-    type: "h3",
-    text: "Your Rights under the GDPR",
-  },
-  {
-    type: "paragraph",
-    text: "The Company undertakes to respect the confidentiality of Your Personal Data and to guarantee You can exercise Your rights.",
-  },
-  {
-    type: "paragraph",
-    text: "You have the right under this Privacy Policy, and by law if You are within the EU, to:",
-  },
-  {
-    type: "list",
-    items: [
-      "Request access to Your Personal Data. The right to access, update or delete the information We have on You. Whenever made possible, you can access, update or request deletion of Your Personal Data directly within Your account settings section. If you are unable to perform these actions yourself, please contact Us to assist You. This also enables You to receive a copy of the Personal Data We hold about You.",
-      "Request correction of the Personal Data that We hold about You. You have the right to have any incomplete or inaccurate information We hold about You corrected.",
-      "Object to processing of Your Personal Data. This right exists where We are relying on a legitimate interest as the legal basis for Our processing and there is something about Your particular situation, which makes You want to object to our processing of Your Personal Data on this ground. You also have the right to object where We are processing Your Personal Data for direct marketing purposes.",
-      "Request erasure of Your Personal Data. You have the right to ask Us to delete or remove Personal Data when there is no good reason for Us to continue processing it.",
-      "Request the transfer of Your Personal Data. We will provide to You, or to a third-party You have chosen, Your Personal Data in a structured, commonly used, machine-readable format. Please note that this right only applies to automated information which You initially provided consent for Us to use or where We used the information to perform a contract with You.",
-      "Withdraw Your consent. You have the right to withdraw Your consent on using your Personal Data. If You withdraw Your consent, We may not be able to provide You with access to certain specific functionalities of the Service.",
-    ],
-  },
-  {
-    type: "h3",
-    text: "Exercising of Your GDPR Data Protection Rights",
-  },
-  {
-    type: "paragraph",
-    text: "You may exercise Your rights of access, rectification, cancellation and opposition by contacting Us. Please note that we may ask You to verify Your identity before responding to such requests. If You make a request, We will try our best to respond to You as soon as possible.",
-  },
-  {
-    type: "paragraph",
-    text: "You have the right to complain to a Data Protection Authority about Our collection and use of Your Personal Data. For more information, if You are in the European Economic Area (EEA), please contact Your local data protection authority in the EEA.",
-  },
-  {
-    type: "h2",
-    text: "Children's Privacy",
-  },
-  {
-    type: "paragraph",
-    text: "Our Service does not address anyone under the age of 13. We do not knowingly collect personally identifiable information from anyone under the age of 13. If You are a parent or guardian and You are aware that Your child has provided Us with Personal Data, please contact Us. If We become aware that We have collected Personal Data from anyone under the age of 13 without verification of parental consent, We take steps to remove that information from Our servers.",
-  },
-  {
-    type: "paragraph",
-    text: "If We need to rely on consent as a legal basis for processing Your information and Your country requires consent from a parent, We may require Your parent's consent before We collect and use that information.",
-  },
-  {
-    type: "h2",
-    text: "Links to Other Websites",
-  },
-  {
-    type: "paragraph",
-    text: "Our Service may contain links to other websites that are not operated by Us. If You click on a third party link, You will be directed to that third party's site. We strongly advise You to review the Privacy Policy of every site You visit.",
-  },
-  {
-    type: "paragraph",
-    text: "We have no control over and assume no responsibility for the content, privacy policies or practices of any third party sites or services.",
-  },
-  {
-    type: "h2",
-    text: "Changes to this Privacy Policy",
-  },
-  {
-    type: "paragraph",
-    text: "We may update Our Privacy Policy from time to time. We will notify You of any changes by posting the new Privacy Policy on this page.",
-  },
-  {
-    type: "paragraph",
-    text: 'We will let You know via email and/or a prominent notice on Our Service, prior to the change becoming effective and update the "Last updated" date at the top of this Privacy Policy.',
-  },
-  {
-    type: "paragraph",
-    text: "You are advised to review this Privacy Policy periodically for any changes. Changes to this Privacy Policy are effective when they are posted on this page.",
-  },
-  {
-    type: "h2",
-    text: "Contact Us",
-  },
-  {
-    type: "paragraph",
-    text: "If you have any questions about this Privacy Policy, You can contact us:",
-  },
-  {
-    type: "paragraph",
-    text: "By email: 3ravens3@gmail.com",
-  },
-];
