--- src/data/writing/before-the-world-dried-up-data.ts
@@ -1,128 +1,0 @@
-import { StoryDataProps } from "@/src/types/data-props";
-
-export const beforeTheWorldDriedUpData: StoryDataProps = {
-  title: "Before the World Dried Up",
-  subtitle: "Shane Bonkowski",
-  date: "June 20, 2024",
-  imageUrl: "/webps/writing/before-the-world-dried-up.webp",
-  imageWidth: 500,
-  imageHeight: 422,
-  body: [
-    {
-      content: `It was a year so far in the distant future that they decided to stop counting altogether. Whether they lost count or it simply didn't matter anymore remains a mystery. Two distant cosmic archaeologists, armed with the goal of tracking the origins of man, find themselves in a long-forgotten corner of some long since forgotten galaxy.
-
-How they ended up here was unclear, but the planet they sighted was scorched beyond recognition. It was a small, rocky planet orbiting a dim White Dwarf with just enough energy to glow, and nothing more. Cleary not a place hospitable for life, but maybe one day years and years ago it would have been ripe for life. Or so the faint biosignatures radiating off of it led them to believe. 
-
-They almost decided against landing altogether. Time and time again, they had stopped to visit a rocky planet with faint biosignatures to no avail. But today was different. Maybe it was because it had been many millions of miles since their last expedition, and they were growing restless. Or perhaps it was some sort of cosmic intervention pointing them to this lonely planet. Regardless, they landed in what appeared to be a long-dried riverbed, evidenced by the long, branching channels resembling the rivers from their home planet. 
-
-Just when they thought they had explored enough to conclude this was in fact just another lifeless rocky planet, one of them had spotted something strange off in the distance: a lonely vessel in surprisingly good shape. This was especially peculiar, considering that everything else in this godforsaken wasteland had been scorched beyond recognition. Its pure white frame stood out against the burnt black landscape like a star in an empty sky. Inside, they found a pristine skeleton clutching some sort of ancient recording mechanism. They pressed play.`,
-      fontStyle: "italic",
-      textAlign: "justify",
-    },
-    {
-      content: `They ask me why I plan on sticking around this place. Why I haven't jumped ship with the rest of them. I let out a sigh as I gaze off across the lifeless desertscape from the helm of my beached vessel. Its dry, relentless blight siphoning every ounce of life that crosses its path.`,
-      fontStyle: "normal",
-      textAlign: "justify",
-    },
-    {
-      content: `Before me lay a boneyard.
-Devoid of life.
-Devoid of movement.
-Devoid of soul.`,
-      fontStyle: "normal",
-      textAlign: "center",
-    },
-    {
-      content: `I remember a time when Angels graced this planet with their delicate songs. Who shimmered and flocked by the thousands in perfect harmony like the stars above. Lighting the way through the deepest, darkest depths no man had ever laid eyes on before. `,
-      fontStyle: "normal",
-      textAlign: "justify",
-    },
-    {
-      content: `A time of mystery and wonder. 
-Sunken cities, lost treasures,
-Fantastic Aquatic Beasts, just out of reach. 
-Waiting to be discovered.`,
-      fontStyle: "normal",
-      textAlign: "center",
-    },
-    {
-      content: `I remember a time when the mighty, unforgiving sea was King. A gentle beast of immense, unrelenting power that was not to be underestimated. Whose mighty, chaotic waves upheld order and prosperity on this planet.`,
-      fontStyle: "normal",
-      textAlign: "justify",
-    },
-    {
-      content: `Gone are those days, they say.
-Or so they say at least.
-As long as I am here to tell the story,
-Those days will live on.`,
-      fontStyle: "normal",
-      textAlign: "center",
-    },
-    {
-      content: `Or maybe it was our lonely Lunar Guardian who was King. Who with each pass commanded the ravenous tides to devour the earth. Waging war amidst the threshold between land and sea. Each pass signaling a new dawn. A new opportunity to begin again.`,
-      fontStyle: "normal",
-      textAlign: "justify",
-    },
-    {
-      content: `Maybe I stick around to pay my respects.
-To watch after the throne.
-Maybe I stick around to admire it all.
-Or maybe to bask in the absurdity of it all.`,
-      fontStyle: "normal",
-      textAlign: "center",
-    },
-    {
-      content: `As the years pass, my memory of it fades. I dust off the thick coating of dust from the windshield and close my eyes, trying to remind myself what it was like to cruise on this vessel. When we rode, not even the mighty sea could hold us back. Tides would part beneath us like the Red Sea. `,
-      fontStyle: "normal",
-      textAlign: "justify",
-    },
-    {
-      content: `No land in sight for miles.
-Just me, this boat, and Infinity.
-What I’d do for one more ride.
-One final trip to the edge of the world.`,
-      fontStyle: "normal",
-      textAlign: "center",
-    },
-    {
-      content: `They say that our poor Stellar Guardian doesn't have much time left. I wipe the sweat off as its fiery fury beats down on my neck. Well, at least in terms of its lifespan. I’ll be long gone before then. For billions of years, this blazing ball of life, destruction, and sustenance has watched over us.`,
-      fontStyle: "normal",
-      textAlign: "justify",
-    },
-    {
-      content: `Now we sit in the wake of its death throes. 
-Watching as its hellish flames grow closer. 
-Ever larger. 
-Ever hotter.`,
-      fontStyle: "normal",
-      textAlign: "center",
-    },
-    {
-      content: `It’s humbling to think that no matter how far we progress, we will always be at the whims of the Celestial Beasts. To think that somewhere in this grand cosmic balance it was decided that everything we hold dearly would have an expiration date.`,
-      fontStyle: "normal",
-      textAlign: "justify",
-    },
-    {
-      content: `Maybe it is by design.
-Why race if there is no finish line?
-Life has so much impact in just a small blip in time.
-I stand in awe at their chaotic beauty.`,
-      fontStyle: "normal",
-      textAlign: "center",
-    },
-    {
-      content: `Maybe I’m old fashioned, but I have no intentions of following the droves of those who plan to escape. To find a new Stellar Guardian, dooming their descendants to this same fate. No, I think I’ll just wait this one out as the universe has intended.`,
-      fontStyle: "normal",
-      textAlign: "justify",
-    },
-    {
-      content: `Sitting here on my long since beached vessel.
-Dreaming of the days that have long since passed.
-What is yet to come for those who never knew.
-The days before the world dried up.
-      `,
-      fontStyle: "normal",
-      textAlign: "center",
-    },
-  ],
-};
