--- src/data/writing/the-moon-data.ts
@@ -1,97 +1,0 @@
-import { StoryDataProps } from "@/src/types/data-props";
-
-export const theMoonData: StoryDataProps = {
-  title: "The Moon",
-  subtitle: "Shane Bonkowski",
-  date: "July 13, 2024",
-  imageUrl: "/webps/writing/the-moon.webp",
-  imageWidth: 500,
-  imageHeight: 422,
-  body: [
-    {
-      content: `Step by step, I trudge on through the long night. My body aches and my bones tremble under the weight of her gaze, a constant reminder of my imprisonment. There is no backward, only forward. It's all I've ever known. 
-
-On the horizon, gently nestled between two colossal towers lay my captor. Ancient beasts of stone and mortar with no apparent entrance or exit. They look as if they have always existed—as if they will exist long after the universe lets out its last whimper and fades into the night.`,
-      fontStyle: "normal",
-      textAlign: "justify",
-    },
-    {
-      content: `And so does she.`,
-      fontStyle: "italic",
-      textAlign: "justify",
-    },
-    {
-      content: `My gaze lay trapped in her orbit. Her look of disappointment taunts me with silent judgment. No matter how hard I try I cannot look away. Nor can I remember how I got here, or what I seek, for that matter. At this point, I carry on to spite her. 
-
-The road I travel is as lonely as it is unending, winding through barren hills and desolate valleys. The landscape stretches out in stark emptiness, with jagged rocks and twisted trees grasping at the empty sky for anything it has to offer. There are no signs of life along its path, no comforting sounds of nature—I am joined only by those who shimmer in the shadows. Manifestations of horror and evil, whispering and laughing. Those who I cannot directly observe, but whose presence haunts me. They laugh and laugh with awful sounds of metal creaking and whining. They whisper of my weakness. The things they would do to me if not for their ruler holding them back. 
-
-Pausing briefly to catch my breath and rest my tired limbs, I sit and ponder for a moment. Drowning out the voices surrounding me, I dwell on questions of existence and purpose—the things I must have done in my past life to have deserved this cruel treatment. My introspection is short-lived, as the voices grow impossible to ignore any longer. They circle all around me with their silent threats, growing closer by the second. Everything in my being tells me to give in to these dark forces and let them devour me. Just before they have the chance to close in and snatch me away, I conjure up the strength to stand. 
-
-“Why do you do this to me?” I shout into the void at my captor. 
-
-I am reminded why I do not usually ask these questions aloud. 
-
-“The answer lies at the end of the road,” the same response as always.
-
-And so, defeated as ever, I embark again. And with that, the shadows retreat to the safety of darkness. My body burns with fatigue and my heart burns with rage. I do not eat. I do not sleep. I only march. 
-
-I've been on this journey as long as I can remember, and no matter how far I push forward, the towers never seem to grow closer. They lurk on the horizon waiting and watching, their colossal forms permanently etched against the starless sky. Their surfaces, worn down by eons of existence, bear silent witness to the countless souls who have sought what lies at the end of the road.
-
-With each step, my silence grows deeper. Many, many paces come and go before I decide to speak again.
-
-“Why won't you just let me die already?” I mutter under my breath, pausing for a brief moment.
-
-Before she gets the chance to reply, the ever-lurking shadows seize upon my moment of vulnerability. Their cold presence surrounds me, squeezing the air from my lungs. They won't be letting me go as easily this time. 
-
-Somehow I find a way to break my gaze and stare into the indiscriminate reflections on their mirror-finished skin. Their grotesque bodies contort and shimmer in the night, making it impossible to make out their form. 
-
-“Why do you torment me?” I cry out.
-
-Their response is immediate and violent. In an instant, I am enveloped by their darkness. Their forms twist and writhe around me, and I am lost in a sea of shadows. Their whispers invade my mind, drowning out my thoughts with promises of pain and despair. They pull at my limbs, their touch like ice, sending a chill through every fiber of my being.
-
-An eternity passes in a blur of torment. I am trapped in a never-ending cycle of agony, each moment more excruciating than the last. My body left battered, my spirit broken. Just when I think I cannot endure another second, I catch a glimpse of my captor. Her gaze still upon me, unyielding as ever. 
-
-“The answer lies at the end of the road,” she speaks softly with a warm embrace.
-
-The shadows retreat, and I am left gasping for breath. My body trembles with immeasurable pain, my mind a broken shell, my soul burdened with despair. 
-
-And there she lay gently suspended against the empty sky, her pale face etched with ancient wisdom and silent observation. Emitting a brilliant radiant light, illuminating the way for this weary traveler. 
-
-“Thank you,” I whisper through the pain, left with no other option but to rise again and resume my unending march.
-
-Another indeterminate amount of time passes. Maybe weeks. Maybe years. Maybe centuries. All this time I have not been able to shake that phrase from my mind, nor their shimmering image from my eyes. My bones ache and whine, my lungs fight for each whimpering breath, and my lips left sealed shut from lack of use. Somehow, I find it in me to force a sentence through.
-
-“Are you God?” 
-
-“The answer lies at the end of the road,” she states, sure as the unending night. 
-
-“What do you want from me?” I plead. 
-
-“The answer lies at the end of the road.”
-
-As I march on, the whispers only grow louder. Silent screams surround me, begging me to give in. Begging for another soul to add to their collection. I have put up with this for so long, but I grow frail and weak. I do not know how much longer I can put up with this torment. 
-
-“Would you stop them again if they came for me?”
-
-“The answer lies–” she begins, but she cannot finish her sentence before my legs give out. My body contorts and spins, and the back of my head slams onto the paved path before me. My eyes still locked onto her like a magnet, my body twisted and knotted. Sparing no time, the shimmering shadows swoop in to envelop me. They creak with their awful grating sounds. Their whispers infest me like termites in an old rotting log. I am fully prepared to be ripped to pieces.
-
-Just as they reach for me, I catch a glimpse of her in the reflection of their shining skin. Peering back at me with a face that has seen it all, she gently winks and nods her head. 
-
-“–at the end of the road,” she concludes. Her look of disappointment now a soft smile.
-
-In an act of desperation and newfound courage, I lunge at one of the abominations, my gaze unbroken on her perfect image. And suddenly, in an awful cascade of broken glass, it explodes into a million pieces before me. And so too the other shadows begin to shatter one by one. I cannot confirm with my eyes, but my ears tell me that the towers followed suit. 
-
-Amidst the magnificent glitter of the rubble, there her image lay before me in perfect form, unbroken and unphased. I sit and admire her for a long while before dozing off for a long-needed rest. 
-
-As I wake, sure enough there they are again. The two towers in all their colossal might. Only this time, a new entity nestled between them. Radiating a blinding light so intense that I have no choice but to cover it up with her image. Everything about this direction is the same as before, except now the long night has passed and she has grown a fiery mane around her. 
-
-I approach the base of the towers as her fire grows fiercer. I am left with no choice but to close my eyes for protection against her blinding light. My strides grow short and careful. Step by step, I trudge on through the inferno. I stretch forward to test the path ahead only to realize there is no more ground beneath me. I open my eyes and let the white blinding light flood in. 
-
-I glance at my mighty ruler for guidance. In her eternal wisdom, she looks at peace with my decision. I cast the shard and watch as it sparkles into infinity below me. 
-
-With a deep breath, I step into the light as the curtains close behind me. With one last glimpse to take it all in, I watch as all that glitters among the rubble ascends to the sky above to keep her company through the long night. Until we meet again.`,
-      fontStyle: "normal",
-      textAlign: "justify",
-    },
-  ],
-};
